+++ drift/lib/src/runtime/query_builder/expressions/expression.dart
@@ -142,13 +142,29 @@ abstract class Expression<D extends Object> implements FunctionParameter {
   /// An expression that is true if `this` resolves to any of the values in
   /// [values].
   Expression<bool> isIn(Iterable<D> values) {
-    return _InExpression(this, values.toList(), false);
+    return isInExp([for (final value in values) Variable<D>(value)]);
   }
 
   /// An expression that is true if `this` does not resolve to any of the values
   /// in [values].
   Expression<bool> isNotIn(Iterable<D> values) {
-    return _InExpression(this, values.toList(), true);
+    return isNotInExp([for (final value in values) Variable<D>(value)]);
+  }
+
+  /// An expression that evaluates to `true` if this expression resolves to a
+  /// value that one of the [expressions] resolve to as well.
+  ///
+  /// For an "is in" comparison with values, use [isIn].
+  Expression<bool> isInExp(List<Expression<D>> expressions) {
+    return _InExpression(this, expressions, false);
+  }
+
+  /// An expression that evaluates to `true` if this expression does not resolve
+  /// to any value that the [expressions] resolve to.
+  ///
+  /// For an "is not in" comparison with values, use [isNotIn].
+  Expression<bool> isNotInExp(List<Expression<D>> expressions) {
+    return _InExpression(this, expressions, true);
   }
 
   /// An expression checking whether `this` is included in any row of the