+++ drift/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.11.0
 
 - Add support for subqueries in the Dart query builder.
+- Add `isInExp` and `isNotInExp` to construct `IS IN` expressions with arbitrary
+  expressions.
 
 ## 2.10.0